Stage intro and results (#114)
* Match sub_802DBC0

* Match sub_802E044

* Match sub_802E0D4

* Match sub_802E164 and sub_802E1EC

* Move sub_802E278 and sub_802E384 into nonmatch

* Start SetupStageIntro

* Progress on SetupStageIntro

* Move SetupStageIntro into nonmatch

* Match Task_802F75C

* Move data to C

* Move Task_802F9F8 into nonmatch

* Match Task_IntroColorAnimation

* Document some of Task_IntroColorAnimation

* Match StageIntroUpdateIcons

* Progress on Task_IntroZoneNameAndIconAnimations

* Move Task_IntroZoneNameAndIconAnimations into nonmatch

* Formatter........

* Move Task_IntroActLettersAnimations into nonmatch

* Match remaining procedures of intro module

* Make most procedures static

* Fix function header types

* Progress on sub_8031138

* Move sub_8031138 into nonmatching

* Match sub_80310F0

* Move Task_UpdateGotThroughScreen into non-matching

* Start sub_80304DC and fix type of StageOutro.unk170

* Introduce SPECIAL_STAGE_REQUIRED_SP_RING_COUNT

* Progress on sub_80304DC

* Finish stage_outro.s

* Rename stage/outro.c -> stage/results.c

* Rename task creation to CreateStageResults

* Use ZONE_TIME_TO_INT in results.c

* Address review issues

+++ src/game/stage/ui.c
@@ -464,12 +464,12 @@ void Task_CreateStageUIMain(void)
 }
 
 // Almost identical to Debug_PrintIntegerAt()"
-void StageUI_PrintIntegerAt(u32 value, u16 x, u16 y, u8 palId)
+void StageUI_PrintIntegerAt(u32 value, s16 x, s16 y, u8 palId)
 {
     StageUI *ui = TaskGetStructPtr(gStageUITask);
     Sprite *digits = &ui->digits[0];
     u32 numDigits;
-    u32 digitX;
+    s32 digitX;
 
     s32 base = 10;
     u32 remaining = 1;
@@ -493,7 +493,8 @@ void StageUI_PrintIntegerAt(u32 value, u16 x, u16 y, u8 palId)
     if (remaining == 0 || numDigits == 0)
         return;
 
-    for (digitX = x; remaining > 0 && numDigits > 0; digitX -= 8, numDigits--) {
+    // NOTE: (u16)x cast needed for matching
+    for (digitX = (u16)x; remaining > 0 && numDigits > 0; digitX -= 8, numDigits--) {
         Sprite *digit;
 
         remaining = Div(value, base);
